Finishes first pass at bitcoin serializer and cleans up organization Untested, but should move all the necessary data around resulting in full serialization of the transaction. Also, the currency-specific utils have been moved to their own files for better organization.
