+++ index.js
@@ -1,6 +1,3 @@
-const Bitcoin = require('./src/providers/Bitcoin');
-const debug = require('debug');
-const Ethereum = require('./src/providers/Ethereum');
 const Client = require('./src/client');
 
 // Client.prototype.tokensBySymbol = require('./lib/tokensBySymbol.json');
@@ -10,8 +7,4 @@ const Client = require('./src/client');
 
 module.exports = {
   Client,
-  Providers: {
-    Bitcoin,
-    Ethereum,
-  }
 };
